requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
gevent==23.9.1
gunicorn==21.2.0
